Add files via upload
A/B tests are very commonly performed by data analysts and data scientists. For this project, you will be working to understand the results of an A/B test run by an e-commerce website. Your goal is to work through this notebook to help the company understand if they should:

    Implement the new webpage,
    Keep the old webpage, or
    Perhaps run the experiment longer to make their decision.

Each ToDo task below has an associated quiz present in the classroom. Though the classroom quizzes are not necessary to complete the project, they help ensure you are on the right track as you work through the project, and you can feel more confident in your final submission meeting the rubric specification.